+++ src/Index.js
@@ -5,8 +5,8 @@ import CreateButton from './CreateButton'
 import {
   getActions,
   getHasIndex,
-  getIndexFields,
-  getModelLabelPlural
+  getIndexFields, getModelLabel,
+  getModelLabelPlural, getSingleton
 } from './utils/schemaGetters'
 import { Redirect } from 'react-router-dom'
 import {
@@ -132,6 +132,30 @@ const Index = ({
   tableOptions,
   customProps
 }) => {
+
+  // if singleton, Index redirects to Detail pg
+  if (getSingleton(schema, modelName)) {
+    const singleton = R.last(data)
+    // singleton may not be null when last deleted; test for 'id'
+    const singleId = R.propOr(null, 'id', singleton)
+    if (singleId) {
+      return <Redirect to={`/${modelName}/${singleId}`} />
+    }
+    // if no singleId exists, must create
+    const actions = getActions(schema, modelName)
+    const onCreateClick = R.path(['create', 'onIndexCreate'], actions)
+    return (
+      <div className='container'>
+        <h1>
+          {`No ${getModelLabel({ schema, modelName, data, user, customProps })} Exists`}
+          <CreateButton {...{
+            onClick: () => onCreateClick({ modelName })
+          }} />
+        </h1>
+      </div>
+    )
+  }
+
   const IndexOverride = getIndexOverride(schema, modelName)
 
   const IndexComponent = IndexOverride || DefaultIndex