+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ### Unreleased
 
+-   Add singleton feature #113
 -   Add disabled options dropdown #111
 
 ## Version 1.4.0